Define simple program to initialize EGL
This sample was copied from the nVidia blog post, that defines
the procedure of how to initialize EGL that is valid for OpenGL
rendering. I used this as an example of how to use the library
so that I can start to define those things.

The original blog post can be found at:
https://developer.nvidia.com/blog/egl-eye-opengl-visualization-without-x-server/

Signed-off-by: Wynand Pieterse <wpieterse2825@gmail.com>